[ML-Python] Loan Risk Prediction Project
This report presents the analysis process using machine learning techniques on the Home Equity Line of Credit (HELOC) application made by real homeowners. The aim of this analysis is to predict the likelihood that applicants will repay their HELOC account within 2 years. This credit scoring method could be used by financial institutions when deciding whether or not to approve a loan. In this report, we did data cleaning, train-test split, and built six predictive models using the train set. Next, we evaluated the model performance based on different performance matrix. Lastly, an interface was built to visualize and simplify the prediction model for users.