[NLP] Airline Twitter Sentiment Classification
In our data, there are 4555 tweets here. Here, we mainly perform sentiment analysis on these 4555 tweets. I will extract negative words from 1,700 complaint tweets, positive words from 1,700 non-complaint tweets, and then use rule-based classification to model to obtain the sentiment analysis prediction results of 4555 tweets data. Finally, the predicted sentiment analysis results are compared with the manually-selected correct sentiment analysis results to calculate the accuracy of the model classification to evaluate the effect of the model.
The rule-based classification can be used to refer to any classification scheme that make use of IF-THEN rules for class prediction. So we first use Corpus () function to create a corpus, then remove the numbers, symbols, stop words, etc. in the corpus to get a meaningful corpus as much as possible, and finally filter the emotions based on the high-frequency words in the corpus. Classification of rule words to predict sentiment of text.