[Case Competition]- Iron Mountain Pricing
Iron Mountain Incorporated (NYSE: IRM), founded in 1951 is the global leader for storage and information management services. 
A segment of IRM that accounts for approximately $80 million of revenue per year recently acquired several entities. The acquired locations may have lower prices to represent the prior lower value.  We need to understand the current pricing for this segment to optimize our Revenue Management Strategy in 2020. 
In the case, we focused on 'LA-Acquired' market and design pricing scheme for different service and storage types of this market. 
 - Set LA market's price as reference, and set mark-ups
 - Re-weight average price based on invoice quantity
 - Explore price difference in different bill code type and es/non-es market respectively
 - Define large, medium, small companies based on their number of child account 
 - Evaluate potential impact of recommendations (advantages, disadvantages, and limitations due to lack of information)