Southwest Conquistador Beer Feasibility Analysis
In this project, I thought in the shoes of Larry Gomez, who are undetermined on whether to acquire a distributorship of Conquistador beer in southwest Oregon. I calculates projected cost, revenue and profit based on actual data, secondary research, and reasonable assumptions. Finally I used NPV analysis to make the decision.