A/B Test on Email Campaign with Causal Forest
A wine retailer is running email marketing experiments to evaluate offers prior to sending the offers to a broader set of customers. Our analytic goal is to evaluate of the average causal effect of email campaign on consumer purchasing of 4 wine brands.
In this case, we conducted slice and dice analysis (define cut-offs based on recency, frequency, purchase amount, segments etc.) to illustrate the potential for targeting on responses for this email campaign. Different performance metric including impression, click through rate, and conversion rate was examined on different consumer segments to identify variables that could most effectively split customers. At the end, we applied causal forest to evaluate “Individual-level” conditional causal effect, and used it along with the margin and cost information to score each customer (projected profit). 

Insights & Recommendation:
- purch_chard and past_purch are the two variable that could most effectively split customers, and the retailer should target chard segment and big volume purchasers.
- Causal forest analysis predicts that 55.79% of the training sample will be targeted by email.
- In order to improve effectiveness of email marketing, our store can optimize email content with wine categories such as Chard.