Pricing with Logit Model and K-means Segmentation
As the product and pricing manager of Kiwi, a major soft drink company I am now undetermined whether to launch kiwi bubble to compete with competitor's product, mango bubble. I have consumer loyalty card data in the test market to aid global launching and pricing decisions.
In this case, we implemented multinomial logit models both with and without segmentation and compared different results. We found that with segmentation, we could achieve a higher profit with higher profit-maximization price because of different own and cross elasticity results. In the 'with segmentation' part, we grouped consumers into segments using "kmeans" function by all columns from the demographic data, then estimated multinomial logit model separately for each segment of consumers.
With segmentation, we also compared between situations with KB and without KB. We found that launching KB is a good movement for Kiwi because of the higher profit and market share.
Lastly, we deeply studied competitor responses by calculating the equilibrium price under Nash equilibrium. Summary of the strategic advantage of Kiwi Bubbles at the ‘Price War’ were provided in the last part of the report.